Autenticação e ajustando as permissões

+++ src/app/controllers/VehiclesController.ts
@@ -4,53 +4,53 @@ import Vehicle from '../models/Vehicle';
 
 class VehiclesController{
     async create(request: Request, response: Response){
+        const userLogged = request.user
+
+        if(userLogged.role !== 'admin'){
+            return response.status(403).json({error: 'Você não tem permissão para adicionar veículos'})
+        }
+
         const {brand, model, year, km, color, chassis, sale_price, cost_price} = request.body
+
+        const dataMandatory = ['brand', 'model', 'year', 'km', 'color', 'chassis', 'sale_price', 'cost_price']
+
+        const errors: Array<any> = []
+
+        dataMandatory.forEach(element => {
+            if(!request.body[element]){
+                errors.push({
+                    field: element,
+                    message: `O campo ${element} é obrigatório`
+                })
+            }
+        })
+
+        if(errors.length > 0){
+            return response.json({errors})
+        }
+
         const data = await VehiclesService.create({brand, model, year, km, color, chassis, sale_price, cost_price})
 
         return data["error"] ? response.status(400).json(data.error) : response.status(201).json(data.data)
     }
 
     async findAll(request: Request, response: Response){
-        const vehicles = await Vehicle.findAll()
-        return response.json(vehicles)
-    }
+        const page = parseInt(request.query.page?.toString() ?? "1")
+        const limit = parseInt(request.query.limit?.toString() ?? "7")
 
-    async findAllPaginated(request: Request, response: Response){
-        const pageNumber: number = parseInt(request.params.pageNumber)
-        let offSet = 0
-
-        if(isNaN(pageNumber)){
-            return response.status(400).json({error: "parâmetro inválido"})
-        }else{
-            offSet = (pageNumber -1) * 7
+        if(isNaN(page) || isNaN(limit)){
+            return response.status(400).json({error: "Parâmetros inválidos"});
         }
 
-        const vehicles = await Vehicle.findAndCountAll({
-            limit: 7,
-            offset: offSet,
-            order: [
-                ['id', 'DESC']
-            ],
-            attributes: {
-                exclude: ['cost_price', 'createdAt', 'updatedAt']
-            }
+        if(page < 1 || limit < 1){ 
+            return response.status(400).json({error: "Parâmetros inválidos"});
+        }
+
+        const vehicles = await Vehicle.findAll({ 
+            limit, 
+            offset: (page - 1) * limit
         })
-
-        let next: boolean
-
-        if(offSet + 7 >= vehicles.count){
-            next = false
-        }else{
-            next = true
-        }
-
-        var result = {
-            pageNum: pageNumber,
-            vehicles: vehicles.rows,
-            next: next
-        }
-
-        return response.status(200).json(result)
+        return response.json(vehicles)
     }
 
     async findByPk(request: Request, response: Response){
@@ -77,6 +77,24 @@ class VehiclesController{
     async update(request: Request, response: Response){
         const id = parseInt(request.params.id)
         const {km, color, sale_price} = request.body
+
+        const dataMandatory = ['km', 'color', 'sale_price'];
+
+        const errors: Array<any> = []
+
+        dataMandatory.forEach(element => {
+            if(!request.body[element]){
+                errors.push({
+                    field: element,
+                    message: `O campo ${element} é obrigatório`
+                })
+            }
+        })
+
+        if(errors.length > 0) {
+            return response.json({ errors })
+        }
+
         const data = await VehiclesService.findByPk(id)
 
         if(data["error"]) return response.status(404).json(data)
@@ -87,6 +105,13 @@ class VehiclesController{
 
     async destroy(request: Request, response: Response){
         const id: number = parseInt(request.params.id)
+
+        const userLogged = request.user;
+
+        if(userLogged.role !== 'admin') {
+            return response.status(403).json({ error: 'Você não tem permissão para remover veículos' });
+        }
+
         const data = await VehiclesService.destroy(id)
 
         return data["error"] ? response.status(400).json(data) : response.status(200).json(data.data)
